Fix MSVC build

+++ src/survive.c
@@ -96,7 +96,7 @@ static void *button_servicer(void * context)
 
 void survive_verify_FLT_size(uint32_t user_size) {
   if(sizeof(FLT) != user_size) {
-    fprintf(stderr, "FLT type incompatible; the shared library libsurvive has FLT size %lu vs user program %u\n", sizeof(FLT), user_size);
+    fprintf(stderr, "FLT type incompatible; the shared library libsurvive has FLT size %lu vs user program %u\n", (unsigned long)sizeof(FLT), user_size);
     fprintf(stderr, "Add '#define FLT %s' before including survive.h or recompile the shared library with the appropriate flag. \n",
 	    sizeof(FLT) == sizeof(double) ? "double" : "float");
     exit(-1);
